+++ app/track.cpp
@@ -138,7 +138,7 @@ bool Track::reload() {
   return fillAudioProperties() && fillTags();
 }
 
-void Track::setDuration(quint32 dur) {
+void Track::setDuration(quint64 dur) {
   _duration = dur;
 }
